update for arduino lib format >=1.5
