+++ library.properties
@@ -1,0 +1,10 @@
+name=collar
+version=1.0.0
+author=CrashOverride85 <crashoverride85@outlook.com>
+maintainer=CrashOverride85 <crashoverride85@outlook.com>
+sentence=Control shock collars
+paragraph=Supports transmitting and receiving commands for shock collars
+category=Communication
+url=https://github.com/CrashOverride85/collar
+architectures=*
+includes=collar.h
